[router]restructure protocol modules for better organization (#9321)
