+++ sgl-router/src/lib.rs
@@ -5,8 +5,8 @@ use std::collections::HashMap;
 pub mod core;
 pub mod metrics;
 pub mod middleware;
-pub mod openai_api_types;
 pub mod policies;
+pub mod protocols;
 pub mod reasoning_parser;
 pub mod routers;
 pub mod server;
