+++ sgl-router/src/protocols/mod.rs
@@ -1,0 +1,6 @@
+// Protocol definitions and validation for various LLM APIs
+// This module provides a structured approach to handling different API protocols
+
+pub mod common;
+pub mod generate;
+pub mod openai;
