+++ sgl-router/src/routers/mod.rs
@@ -9,7 +9,10 @@ use axum::{
 };
 use std::fmt::Debug;
 
-use crate::openai_api_types::{ChatCompletionRequest, CompletionRequest, GenerateRequest};
+use crate::protocols::{
+    generate::GenerateRequest,
+    openai::{chat::ChatCompletionRequest, completions::CompletionRequest},
+};
 
 pub mod factory;
 pub mod header_utils;
