+++ sgl-router/tests/benchmark_integration.rs
@@ -5,9 +5,13 @@
 
 use serde_json::{from_str, to_string, to_value};
 use sglang_router_rs::core::{BasicWorker, WorkerType};
-use sglang_router_rs::openai_api_types::{
-    ChatCompletionRequest, ChatMessage, CompletionRequest, GenerateParameters, GenerateRequest,
-    SamplingParams, StringOrArray, UserMessageContent,
+use sglang_router_rs::protocols::{
+    common::StringOrArray,
+    generate::{GenerateParameters, GenerateRequest, SamplingParams},
+    openai::{
+        chat::{ChatCompletionRequest, ChatMessage, UserMessageContent},
+        completions::CompletionRequest,
+    },
 };
 
 /// Create a default GenerateRequest for benchmarks with minimal fields set
